+++ escaperoom_main.py
@@ -35,6 +35,7 @@ import random
 from adventurelib import Room, when, say, start, Bag, Item, set_context
 import adventurelib
 import sys
+from inventory import *
 from room_1 import *
 from room_2 import *
 from room_3 import *
@@ -112,30 +113,6 @@ mit breitem IT-Wissen - denn du hast ja DACS studiert ;) - nach einer
 Lösung."""
 )
 
-
-
-# Inventar #
-crowbar = Item("brecheisen")
-smartphone = Item("smartphone")
-hairpin = Item("haarnadel")
-sim = Item("simkarte")
-inventory = Bag()
-
-
-@when("inventar")
-@when("inventar zeigen")  # zeigen
-@when("zeige inventar")
-@when("inventar anzeigen")  # anzeigen
-@when("anzeige vom inventar")
-@when("öffne inventar")  # öffnen
-@when("inventar öffnen")
-def zeige_inventar():
-    print("Du hast: ")
-    if not inventory:
-        print("nichts")
-        return
-    for item in inventory:
-        print(f"*{item}")
 
 #Start
 set_context("room1")
